Move DC Nav logo to general assets folder
